# internal/delivery/api/docs/shipper/shipper_swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Shippers API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/loads": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List loads with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loads"
                ],
                "summary": "List loads",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID filter",
                        "name": "company_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Carrier ID filter",
                        "name": "carrier_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/query.LoadResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new load",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loads"
                ],
                "summary": "Create load",
                "parameters": [
                    {
                        "description": "Create Load Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_karavanix_karavantrack-api-server_internal_usecase_loads_command.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_karavanix_karavantrack-api-server_internal_usecase_loads_command.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    }
                }
            }
        },
...
        "/loads/{id}/assign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assign load to a carrier and truck",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loads"
                ],
                "summary": "Assign load",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Load ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Assign Load Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/command.AssignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    }
                }
            }
        },
        "/loads/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel a load",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loads"
                ],
                "summary": "Cancel load",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Load ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    }
                }
            }
        },
        "/loads/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Confirm a completed load delivery (by company/dispatcher)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loads"
                ],
                "summary": "Confirm load delivery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Load ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    }
                }
            }
        },
        "/loads/{id}/position": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current tracked position for a load (latest GPS point)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loads"
                ],
                "summary": "Get current position",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Load ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/query.PositionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    }
                }
            }
        },
        "/loads/{id}/track": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get location history tracking for a load",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loads"
                ],
                "summary": "Get load track",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Load ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max number of points (default 100, max 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/query.GetTrackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    }
                }
            }
        },
        "/users/carriers/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search for carrier users by name, email, or phone",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Search carriers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query (name, email, or phone)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/query.UsersResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/outerr.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "command.AssignRequest": {
            "type": "object",
            "required": [
                "carrier_id"
            ],
            "properties": {
                "carrier_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_karavanix_karavantrack-api-server_internal_usecase_loads_command.CreateRequest": {
            "type": "object",
            "required": [
                "company_id",
                "dropoff_lat",
                "dropoff_lng",
                "pickup_lat",
                "pickup_lng",
                "title"
            ],
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dropoff_address": {
                    "type": "string"
                },
                "dropoff_lat": {
                    "type": "number"
                },
                "dropoff_lng": {
                    "type": "number"
                },
                "pickup_address": {
                    "type": "string"
                },
                "pickup_lat": {
                    "type": "number"
                },
                "pickup_lng": {
                    "type": "number"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                }
            }
        },
        "github_com_karavanix_karavantrack-api-server_internal_usecase_loads_command.CreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "outerr.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "query.GetTrackResponse": {
            "type": "object",
            "properties": {
                "load_id": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/query.TrackPointResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "query.LoadResponse": {
            "type": "object",
            "properties": {
                "carrier_id": {
                    "type": "string"
                },
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dropoff_address": {
                    "type": "string"
                },
                "dropoff_at": {
                    "type": "string"
                },
                "dropoff_lat": {
                    "type": "number"
                },
                "dropoff_lng": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "member_id": {
                    "type": "string"
                },
                "pickup_address": {
                    "type": "string"
                },
                "pickup_at": {
                    "type": "string"
                },
                "pickup_lat": {
                    "type": "number"
                },
                "pickup_lng": {
                    "type": "number"
                },
                "reference_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "query.PositionResponse": {
            "type": "object",
            "properties": {
                "accuracy_m": {
                    "type": "number"
                },
                "carrier_id": {
                    "type": "string"
                },
                "heading_deg": {
                    "type": "number"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "load_id": {
                    "type": "string"
                },
                "recorded_at": {
                    "type": "string"
                },
                "speed_mps": {
                    "type": "number"
                }
            }
        },
        "query.TrackPointResponse": {
            "type": "object",
            "properties": {
                "accuracy_m": {
                    "type": "number"
                },
                "heading_deg": {
                    "type": "number"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "recorded_at": {
                    "type": "string"
                },
                "speed_mps": {
                    "type": "number"
                }
            }
        },
        "query.UsersResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "API Токен используется для авторизации",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}